Sync with ballerina master

+++ compiler/BIRReader.cpp
@@ -175,7 +175,7 @@ int64_t ConstantPoolSet::getIntCp(int32_t index) {
 }
 
 // Search float from the constant pool based on index
-float ConstantPoolSet::getFloatCp(int32_t index) {
+double ConstantPoolSet::getFloatCp(int32_t index) {
     ConstantPoolEntry *poolEntry = getEntry(index);
     assert(poolEntry->getTag() == ConstantPoolEntry::tagEnum::TAG_ENUM_CP_ENTRY_FLOAT);
     FloatCpInfo *floatCp = static_cast<FloatCpInfo *>(poolEntry);
